requirements.txt
pydantic>=2.7.0
python-telegram-bot>=21.0
openai>=1.0.0
PyYAML>=6.0
keyring>=25.0.0
pytest>=8.0.0
